refactor: split game session execution modules
